gameshow/src/python/video_captioner/requirements.txt
openai-whisper>=20231117
--extra-index-url https://download.pytorch.org/whl/rocm5.7
torch>=2.0.0+rocm5.7
numpy>=1.20.0
tqdm>=4.65.0
more-itertools>=9.1.0
transformers>=4.30.0
ffmpeg-python>=0.2.0 